Backend/requirements.txt
Flask == 2.3.3
SQLAlchemy == 2.0.20
Flask-Migrate == 4.0.4
Flask-SqlALchemy == 3.0.5
Flask-JWT-Extended == 4.4.4
Flask-Cors == 4.0.1
Flask-Bcrypt == 1.0.1
python-dotenv == 1.0.0
validate-email == 1.3